# openapi.yaml
...
components:
  schemas:
...
    AnalysisProcess:
      description: 'An analysis process is a process that describes how one or more
        Files, from a

        Study, are transformed to another set of Files via a Workflow. The analysis

        process also keeps track of the workflow metadata and the Agent that is running

        the Analysis.'
      properties:
        creation_date:
          description: Timestamp (in ISO 8601 format) when the entity was created.
          title: Creation Date
          type: string
        has_agent:
          anyOf:
          - type: string
          - $ref: '#/components/schemas/Agent'
          description: The Agent - a software, institution, or human - that is             executing
            or responsible for executing the workflow.
          title: Has Agent
        has_input:
          anyOf:
          - items:
              type: string
            type: array
          - items:
              $ref: '#/components/schemas/File'
            type: array
          description: The input data File entities used in the Analysis Process.
          title: Has Input
        has_output:
          anyOf:
          - items:
              type: string
            type: array
          - items:
              $ref: '#/components/schemas/File'
            type: array
          description: The output data File entities generated by the Analysis Process.
          title: Has Output
        has_workflow_step:
          allOf:
          - $ref: '#/components/schemas/WorkflowStep'
          description: Workflow Step entity that performs a set of operations on the             input
            data Files to generate a set of output data Files.
          title: Has Workflow Step
        id:
          description: An identifier that uniquely represents an entity.
          title: Id
          type: string
        title:
          description: The title that describes an entity.
          title: Title
          type: string
        update_date:
          description: Timestamp (in ISO 8601 format) when the entity was updated.
          title: Update Date
          type: string
        xref:
          description: Alternate identifiers for an entity.
          items:
            type: string
          title: Xref
          type: array
      title: AnalysisProcess
      type: object
...
    DataAccessPolicy:
      description: 'A Data Access Policy specifies under which circumstances, legal
        or otherwise,

        a user can have access to one or more Datasets belonging to one or more Studies.'
      properties:
        accession:
          description: A unique identifier assigned to an entity for the sole             purpose
            of referring to that entity in a global scope.
          title: Accession
          type: string
        creation_date:
          description: Timestamp (in ISO 8601 format) when the entity was created.
          title: Creation Date
          type: string
        description:
          description: A short description for the Data Access Policy.
          title: Description
          type: string
        has_data_access_committee:
          anyOf:
          - type: string
          - $ref: '#/components/schemas/DataAccessCommittee'
          description: The Data Access Committee linked to this policy.
          title: Has Data Access Committee
        has_data_use_condition:
          anyOf:
          - type: string
          - items:
              $ref: '#/components/schemas/DataUseCondition'
            type: array
          description: Data Use Condition entities that are associated with the             Data
            Access Policy.
          title: Has Data Use Condition
        id:
          description: An identifier that uniquely represents an entity.
          title: Id
          type: string
        name:
          description: The name for an entity.
          title: Name
          type: string
        policy_text:
          description: The terms of data use and policy verbiage should be             captured
            here.
          title: Policy Text
          type: string
        policy_url:
          description: URL for the policy, if available. This is useful if the             terms
            of the policy is made available online at a resolvable URL.
          title: Policy Url
          type: string
        update_date:
          description: Timestamp (in ISO 8601 format) when the entity was updated.
          title: Update Date
          type: string
        xref:
          description: Alternate identifiers for an entity.
          items:
            type: string
          title: Xref
          type: array
      title: DataAccessPolicy
      type: object
...
    Experiment:
      description: 'An experiment is an investigation that consists of a coordinated
        set of actions

        and observations designed to generate data with the goal of verifying, falsifying,

        or establishing the validity of a hypothesis.'
      properties:
        accession:
          description: A unique identifier assigned to an entity for the sole purpose             of
            referring to that entity in a global scope.
          title: Accession
          type: string
        biological_replicates:
          description: A biological replicate is a replicate role that consists of             independent
            biological replicates made from different             individual biosamples.
          title: Biological Replicates
          type: string
        creation_date:
          description: Timestamp (in ISO 8601 format) when the entity was created.
          title: Creation Date
          type: string
        description:
          description: A detailed description of the Experiment.
          title: Description
          type: string
        experimental_replicates:
          description: The replicate number of the assay, i.e. the numeric iteration             for
            the assay that was repeated.
          title: Experimental Replicates
          type: string
        has_experiment_process:
          anyOf:
          - items:
              type: string
            type: array
          - items:
              $ref: '#/components/schemas/ExperimentProcess'
            type: array
          description: One or more Experiment Processes entities associated with             this
            Experiment.
          title: Has Experiment Process
        has_file:
          anyOf:
          - items:
              type: string
            type: array
          - items:
              $ref: '#/components/schemas/File'
            type: array
          description: One or more Files entities that are generated as output of             this
            Experiment.
          title: Has File
        has_sample:
          anyOf:
          - type: string
          - $ref: '#/components/schemas/Sample'
          description: The Sample entity associated with this Experiment.
          title: Has Sample
        has_study:
          anyOf:
          - type: string
          - $ref: '#/components/schemas/Study'
          description: The Study entity associated with this Experiment.
          title: Has Study
        has_technology:
          anyOf:
          - type: string
          - $ref: '#/components/schemas/Technology'
          description: The Technology entity associated with this Experiment.
          title: Has Technology
        id:
          description: An identifier that uniquely represents an entity.
          title: Id
          type: string
        technical_replicates:
          description: A technical replicate is a replicate role where the same             BioSample
            is use e.g.  the same pool of RNA used to assess technical             (as
            opposed to biological)  variation within an experiment.
          title: Technical Replicates
          type: string
        title:
          description: 'Name for the experiment (eg: GHGAE_PBMC_RNAseq).'
          title: Title
          type: string
        type:
          description: The type of Experiment.
          title: Type
          type: string
        update_date:
          description: Timestamp (in ISO 8601 format) when the entity was updated.
          title: Update Date
          type: string
        xref:
          description: Alternate identifiers for an entity.
          items:
            type: string
          title: Xref
          type: array
      title: Experiment
      type: object
    ExperimentProcess:
      description: 'An Experiment Process is a process that describes how a Sample
        is transformed to

        a File via an assay. The Experiment Process also keeps track of the Protocol

        used and the Agent that is running the experiment.'
      properties:
        creation_date:
          description: Timestamp (in ISO 8601 format) when the entity was created.
          title: Creation Date
          type: string
        has_agent:
          anyOf:
          - type: string
          - $ref: '#/components/schemas/Agent'
          description: The Agent - a software, institution, or human - that is             executing
            or responsible for executing the Experiment Process.
          title: Has Agent
        has_input:
          anyOf:
          - type: string
          - $ref: '#/components/schemas/Sample'
          description: The input to the Experiment Process. Usually a Sample entity.
          title: Has Input
        has_output:
          anyOf:
          - type: string
          - $ref: '#/components/schemas/File'
          description: The output of this Experiment Process. Usually a File entity.
          title: Has Output
        has_protocol:
          anyOf:
          - type: string
          - $ref: '#/components/schemas/Protocol'
          description: The Protocol entity used by this Experiment Process.
          title: Has Protocol
        id:
          description: An identifier that uniquely represents an entity.
          title: Id
          type: string
        title:
          description: 'A descriptive title that explains the step(s) involved in             performing
            the experiment leading up to the sequencing of the sample             and
            generation of raw data from the instrument.             (eg: Sample extraction
            -> Target Enrichment) '
          title: Title
          type: string
        update_date:
          description: Timestamp (in ISO 8601 format) when the entity was updated.
          title: Update Date
          type: string
        xref:
          description: Alternate identifiers for an entity.
          items:
            type: string
          title: Xref
          type: array
      title: ExperimentProcess
      type: object
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
...
    Publication:
      description: 'The Publication entity represents a publication. While a publication
        can be any

        article that is published, the minimum expectation is that the publication

        has a valid DOI.'
      properties:
        abstract:
          description: The study abstract that describes the goals.  Can also hold             abstract
            from a publication related to this study
          title: Abstract
          type: string
        creation_date:
          description: Timestamp (in ISO 8601 format) when the entity was created.
          title: Creation Date
          type: string
        id:
          description: A PMID or DOI for the Publication.
          title: Id
          type: string
        title:
          description: The title for the Publication.
          title: Title
          type: string
        update_date:
          description: Timestamp (in ISO 8601 format) when the entity was updated.
          title: Update Date
          type: string
        xref:
          description: One or more cross-references for this Publication.
          items:
            type: string
          title: Xref
          type: array
      title: Publication
      type: object
...
info:
  title: FastAPI
  version: 0.1.0
openapi: 3.0.2
paths:
...
  /analysis_process/{analysis_process_id}:
    get:
      description: Given an AnalysisProcess ID, get the AnalysisProcess record from
        the metadata store.
      operationId: get_analysis_processes_analysis_process__analysis_process_id__get
      parameters:
      - in: path
        name: analysis_process_id
        required: true
        schema:
          title: Analysis Process Id
          type: string
      - in: query
        name: embedded
        required: false
        schema:
          default: false
          title: Embedded
          type: boolean
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalysisProcess'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get an AnalysisProcess
...
  /data_access_policies/{data_access_policy_id}:
    get:
      description: Given a DataAccessPolicy ID, get the DataAccessPolicy record from
        the metadata store.
      operationId: get_data_access_policies_data_access_policies__data_access_policy_id__get
      parameters:
      - in: path
        name: data_access_policy_id
        required: true
        schema:
          title: Data Access Policy Id
          type: string
      - in: query
        name: embedded
        required: false
        schema:
          default: false
          title: Embedded
          type: boolean
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataAccessPolicy'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get a DataAccessPolicy
...
  /experiment_processes/{experiment_process_id}:
    get:
      description: Given a ExperimentProcess ID, get the ExperimentProcess record
        from the metadata store.
      operationId: get_experiment_processes_experiment_processes__experiment_process_id__get
      parameters:
      - in: path
        name: experiment_process_id
        required: true
        schema:
          title: Experiment Process Id
          type: string
      - in: query
        name: embedded
        required: false
        schema:
          default: false
          title: Embedded
          type: boolean
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExperimentProcess'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get a ExperimentProcess
  /experiments/{experiment_id}:
    get:
      description: Given a Experiment ID, get the Experiment record from the metadata
        store.
      operationId: get_experiments_experiments__experiment_id__get
      parameters:
      - in: path
        name: experiment_id
        required: true
        schema:
          title: Experiment Id
          type: string
      - in: query
        name: embedded
        required: false
        schema:
          default: false
          title: Embedded
          type: boolean
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Experiment'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get an Experiment
  /files/{file_id}:
    get:
      description: Given a File ID, get the File record from the metadata store.
      operationId: get_files_files__file_id__get
      parameters:
      - in: path
        name: file_id
        required: true
        schema:
          title: File Id
          type: string
      - in: query
        name: embedded
        required: false
        schema:
          default: false
          title: Embedded
          type: boolean
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/File'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get a File
  /individuals/{individual_id}:
    get:
      description: Given a Individual ID, get the Individual record from the metadata
        store.
      operationId: get_individuals_individuals__individual_id__get
      parameters:
      - in: path
        name: individual_id
        required: true
        schema:
          title: Individual Id
          type: string
      - in: query
        name: embedded
        required: false
        schema:
          default: false
          title: Embedded
          type: boolean
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Individual'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get a Individual
...
  /projects/{project_id}:
    get:
      description: Given a Project ID, get the Project record from the metadata store.
      operationId: get_projects_projects__project_id__get
      parameters:
      - in: path
        name: project_id
        required: true
        schema:
          title: Project Id
          type: string
      - in: query
        name: embedded
        required: false
        schema:
          default: false
          title: Embedded
          type: boolean
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get a Project
...
  /publications/{publication_id}:
    get:
      description: Given a Publication ID, get the Publication record from the metadata
        store.
      operationId: get_publications_publications__publication_id__get
      parameters:
      - in: path
        name: publication_id
        required: true
        schema:
          title: Publication Id
          type: string
      - in: query
        name: embedded
        required: false
        schema:
          default: false
          title: Embedded
          type: boolean
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Publication'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get a Publication
  /samples/{sample_id}:
    get:
      description: Given a Sample ID, get the Sample record from the metadata store.
      operationId: get_samples_samples__sample_id__get
      parameters:
      - in: path
        name: sample_id
        required: true
        schema:
          title: Sample Id
          type: string
      - in: query
        name: embedded
        required: false
        schema:
          default: false
          title: Embedded
          type: boolean
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Sample'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get a Sample
...